+++ assets/src/js/dashboard.js
@@ -37,10 +37,15 @@ document.addEventListener('keydown', function (evt) {
 })
 
 // every 61 seconds without mouse activity, reload the page (but only if tab is active)
-function reloadIfActive() {
-  if (!document.hidden) window.location.reload();
-}
 var reloadTimeout = window.setTimeout(reloadIfActive, 61000);
+function reloadIfActive() {
+  if (!document.hidden) {
+    window.location.reload();
+  } else {
+    // if document hidden, try again in 61s
+    reloadTimeout = window.setTimeout(reloadIfActive, 61000);
+  }
+}
 document.addEventListener('mouseover', function() {
   window.clearTimeout(reloadTimeout);
   reloadTimeout = window.setTimeout(reloadIfActive, 61000);